requirements.txt
flask>=2.0.1
numpy>=1.21.0
pillow>=8.3.1
matplotlib>=3.4.2
torch>=1.9.0
torchvision>=0.10.0
onnxruntime>=1.8.1
opencv-python>=4.5.3
requests>=2.26.0
tqdm>=4.62.1
transformers>=4.20.0
timm>=0.6.12
huggingface-hub>=0.10.1
safetensors>=0.3.0